+++ packages/sdk-core/hardhat.config.ts
@@ -38,6 +38,11 @@ const config: HardhatUserConfig = {
                 : [],
             chainId: 56,
         },
+        mainnet: {
+            url: process.env.RPC_URL || "",
+            accounts: process.env.PRIVATE_KEY ? [process.env.PRIVATE_KEY]: [],
+            chainId: 1,
+        }
     },
     mocha: {
         timeout: 250000,