+++ packages/sdk-core/CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 - Added `downgradeTo` function on `SuperToken` class
+- Added toga events to be part of query
 
 ## [0.5.9] - 2022-12-05