+++ packages/sdk-core/CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.5.9] - 2022-12-15
+
+### Fixes
+- Added CFAForwarderV1 address to sdk-core
+
 ## [0.5.8] - 2022-11-15
 
 ### Added